requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiomultiprocess==0.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-property==0.2.2
async-timeout==5.0.1
attrs==24.2.0
backoff==2.2.1
base58==2.1.1
bittensor==8.4.5
bittensor-cli==8.4.1
bittensor-wallet==2.1.3
bt-decode==0.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cytoolz==1.0.0
decorator==5.1.1
distro==1.9.0
docker-pycreds==0.4.0
ecdsa==0.19.0
eth-hash==0.7.0
eth-keys==0.6.0
eth-typing==5.0.1
eth-utils==2.2.2
exceptiongroup==1.2.2
fastapi==0.110.3
frozenlist==1.5.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
Levenshtein==0.26.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy-opentensor==0.5.0
multidict==6.1.0
munch==2.5.0
nest-asyncio==1.6.0
netaddr==1.3.0
nltk==3.9.1
numpy==2.0.2
numpydantic==1.6.5
openai==1.57.1
packaging==24.2
password-strength==0.0.3.post2
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
py==1.11.0
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.2.0
py-sr25519-bindings==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyNaCl==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-statemachine==2.5.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
retry==0.9.2
rich==13.9.4
scalecodec==1.2.11
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
substrate-interface==1.7.11
sympy==1.13.3
tabulate==0.9.0
termcolor==2.5.0
tiktoken==0.8.0
toml==0.10.0
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wandb==0.19.0
websocket-client==1.8.0
websockets==14.1
xxhash==3.5.0
yarl==1.18.3
